Correlator Update
Cyclic-Prefix:
Cyclic-prefix is a standalone function which copies the end portion of the symbol to its front, this function takes the pointer to the start of the prefix as an input parameter.

Coarse Search remains in time domain, the feasability of the frequency domain methods is going to be assesed during or after audio driver release.

Flipping the prefix was uneccesary, this patch rectifies this bug.

Now the correlator output on the Rx datablock is compared against the object's threshold. A function needs to be implemented to either set this level by listening to the noise when nothing is being transmitted or switch the logic to peak detection.

Added cyclic prefix unit test and updated other unit and integration tests to reflect all changes.

Band-Pass Modulator:
Now band pass modulator & demodulator takes the pointer to the start of the symbol.

FFT Object:
Improved the encoding process by adding function which uses initalized fftw plan to save output to specified destination. This removes the need to copy the whole symbol, which was  was done when symbol was upsampled in band pass modulator.

+++ test/unit/FourierTransformsTest.cpp
@@ -52,8 +52,8 @@ BOOST_AUTO_TEST_CASE(IFFTtoFFT)
     // Generate random floats 
     for (uint16_t i = 0; i < nPoints; i++)
     {
-        ifft.in[i][0] = (float) rand()/RAND_MAX;
-        ifft.in[i][1] = (float) rand()/RAND_MAX;
+        ifft.in[i][0] = (double) rand()/RAND_MAX;
+        ifft.in[i][1] = (double) rand()/RAND_MAX;
     }
     
     // Measure wall time of the ifft execution.
@@ -244,8 +244,8 @@ BOOST_AUTO_TEST_CASE(FFTtoIFFT)
     // Generate random floats 
     for (uint16_t i = 0; i < nPoints; i++)
     {
-        forwardfft.in[i][0] = (float) rand()/RAND_MAX;
-        forwardfft.in[i][1] = (float) rand()/RAND_MAX;
+        forwardfft.in[i][0] = (double) rand()/RAND_MAX;
+        forwardfft.in[i][1] = (double) rand()/RAND_MAX;
     }
     
     // Measure wall time of the ifft execution.
@@ -450,8 +450,8 @@ BOOST_AUTO_TEST_CASE(Reconfiguration)
 
         for (uint16_t i = 0; i < PowOf2Arr[j]; i++)
         {
-            forwardfft.in[i][0] = (float) rand()/RAND_MAX;
-            forwardfft.in[i][1] = (float) rand()/RAND_MAX;
+            forwardfft.in[i][0] = (double) rand()/RAND_MAX;
+            forwardfft.in[i][1] = (double) rand()/RAND_MAX;
         }
 
         // Measure wall time of the ifft execution.
